+++ lib/highlight_text.dart
@@ -4,6 +4,7 @@ import 'package:flutter/services.dart';
 
 export 'src/HighlightedWord.dart';
 export 'src/TextHighlight.dart';
+export 'src/HighlightMap.dart';
 
 class HighlightText {
   static const MethodChannel _channel = const MethodChannel('highlight_text');